Transfer functions from main to helper
