+++ automation-english/main.py
@@ -1,14 +1,8 @@
 import os
-import shutil
+# import shutil
 import sys
-import subprocess
-from pprint import pprint
-from datetime import datetime
-import pypandoc
-import yaml
-import validators
 import english_release
-# import helper
+import helper
 
 ##### Global Vars #####
 
@@ -20,101 +14,6 @@ test_dir = "test_env"
 current_dir = "./"
 included_wgs = []
 
-def copy_file(source_filepath, dest_path):
-
-    if os.path.isfile(dest_path):
-        print(f"Warning: File with same name already exists at destination: {source_filepath}")
-    else:
-        try:
-            print(f"\nCopying: {source_filepath}")
-            shutil.copy2(source_filepath, dest_path)
-            print(f"Copied Successfully")
-        except PermissionError:
-            print("Error: Permission denied.")
-            sys.exit(1)
-        except:
-            print("Error: Unable to copy file.")
-            sys.exit(1)
-
-def copy_dir_files(source_folder_path, dest_folder_path):
-
-    try:
-        if os.path.isdir(source_folder_path):
-            files = os.listdir(source_folder_path)
-            for fname in files:
-                if os.path.isfile(os.path.join(dest_folder_path, fname)):
-                    print(f"File with same name already exists at destination: {os.path.join(source_folder_path, fname)}")
-                else:
-                    copy_file(os.path.join(source_folder_path, fname), dest_folder_path)
-        else:
-            print(f"Warning: Source directory does not exist: {source_folder_path}")
-    except NotADirectoryError:
-        print(f"Error: Source path is not a directory :{source_folder_path}")
-        sys.exit(1)
-    except:
-        print(f"Error: Unable to list files in :{source_folder_path}")
-        sys.exit(1)
-
-def convert_md2tex(md_filename, latex_filename):
-
-    print(f"Converting {md_filename} file to LaTeX")
-    output = pypandoc.convert_file(md_filename, 'latex', outputfile=latex_filename, extra_args=['-f', 'gfm'])
-    assert output == ""
-    print(f"Created successfully: {latex_filename}")
-
-def clean_directory(folder_path):
-    try:
-        if os.path.isdir(folder_path):
-            print(f"\nCleaning directory: {folder_path}")
-            shutil.rmtree(folder_path)
-            print("Directory cleaned Successfully")
-    except:
-        print(f"\nWarning: Unable to clean directory: {folder_path}")
-
-def load_yaml(file_path):
-
-    try:
-        with open(file_path) as stream:
-            data = yaml.safe_load(stream)
-            pprint(data)
-            return data
-    except yaml.YAMLError as exc:
-        print(exc)
-        print("Error: Unable to load data from YAML file.")
-        sys.exit(1)
-
-def decrease_level(metric_path):
-
-    try:
-        print(f"\nDecreasing heading levels by 2 in metric: {metric_path}")
-        cmd = 'sed -i "s/^\#/###/g" ' + metric_path
-        os.system(cmd)
-    except:
-        print(f"Error: Unable to decrease heading levels in metric: {metric_path}.\nMake sure the metric follows the template.")
-        sys.exit(1)
-
-def delete_dictkey(key, dictionary):
-
-    if key in dictionary:
-        del dictionary[key]
-    else:
-        print(f"Warning: Key- {key} not found in {dictionary}")
-
-def is_url(string):
-
-    if validators.url(string):
-        return True
-    else:
-        return False
-
-def clone_repo(url, name, branch):
-
-    try:
-        subprocess.check_call(['git', 'clone', '-b', branch, url, name])
-    except:
-        print(f"Error: Unable to clone/checkout repository from {url}")
-        print("Verify the repository details specified in YAML file.")
-        sys.exit(1)
 
 def language_input():
 
@@ -133,78 +32,6 @@ def language_input():
 
     return choice
 
-def add_front_matter(yaml_data):
-
-    # Create and include front matter files
-    with open("front-matter.tex", "w") as front_matter:
-        if yaml_data["front-matter"] is not None:
-            for page in yaml_data["front-matter"]:
-                if is_url(page):
-                    print(f"\nDownloading file: {page}")
-                    os.system(f"wget {page}")
-                    filename = os.path.basename(page)
-                    name, extension = os.path.splitext(filename)
-                    if extension == ".md":
-                        convert_md2tex(filename, name+".tex")
-                        front_matter.write(f"\input{{{name}}} \n")
-                    elif extension == ".tex":
-                        front_matter.write(f"\input{{{name}}} \n")
-                    else:
-                        print(f"Error: Could not incorporate {page} in front-matter.\nPlease make sure that the URL is valid. Only Markdown/LaTeX file format is supported.")
-                        sys.exit(1)
-                elif os.path.splitext(page)[1] == ".md":
-                    convert_md2tex(page, os.path.splitext(page)[0]+".tex")
-                    front_matter.write(f"\input{{{os.path.splitext(page)[0]}}} \n")
-                elif os.path.splitext(page)[1] == ".tex":
-                    front_matter.write(f"\input{{{os.path.splitext(page)[0]}}}"+"\n")
-                else:
-                    print(f"Error: Could not incorporate {page} in front-matter.\nPlease make sure that the filename is valid. Only Markdown/LaTeX file format is supported.")
-                    sys.exit(1)
-        else:
-            print("Warning: No documents detected for the front-matter")
-
-    with open(master_file_path, "a") as master_file:
-        master_file.write("\n\include{front-matter}")
-
-def add_end_matter(yaml_data):
-
-    # Create and include end matter files
-    with open("end-matter.tex", "w") as end_matter:
-        if yaml_data["end-matter"] is not None:
-            for page in yaml_data["end-matter"]:
-                if is_url(page):
-                    print(f"\nDownloading file: {page}")
-                    os.system(f"wget {page}")
-                    filename = os.path.basename(page)
-                    name, extension = os.path.splitext(filename)
-                    if name == "LICENSE":
-                        os.rename("LICENSE", "LICENSE.md")
-                        convert_md2tex("LICENSE.md", "LICENSE.tex")
-                        end_matter.write("\clearpage\n\section{LICENSE}\n\input{LICENSE}\n")
-                    elif extension == ".md":
-                        convert_md2tex(filename, name+".tex")
-                        end_matter.write(f"\input{{{name}}} \n")
-                    elif extension == ".tex":
-                        end_matter.write(f"\input{{{name}}} \n")
-                    else:
-                        print(f"Error: Could not incorporate {page} in end matter.\nPlease make sure that the URL is valid. Only Markdown/LaTeX file format is supported.")
-                        sys.exit(1)
-
-                elif os.path.splitext(page)[1] == ".md":
-                    convert_md2tex(page, os.path.splitext(page)[0]+".tex")
-                    end_matter.write(f"\input{{{os.path.splitext(page)[0]}}} \n")
-                elif os.path.splitext(page)[1] == ".tex":
-                    end_matter.write(f"\input{{{os.path.splitext(page)[0]}}}"+"\n")
-                elif os.path.splitext(page)[0] == "LICENSE" and os.path.splitext(page)[1] == "":
-                    os.rename("LICENSE", "LICENSE.md")
-                    convert_md2tex("LICENSE.md", "LICENSE.tex")
-                    end_matter.write("\clearpage\n\section{LICENSE}\n\input{LICENSE}\n")
-                else:
-                    print(f"Error: Could not incorporate {page} in end matter.\nPlease make sure that the filename is valid. Only Markdown/LaTeX file format is supported.")
-                    sys.exit(1)
-
-        else:
-            print("Warning: No documents detected for the end-matter")
 
 def main():
 
@@ -215,7 +42,7 @@ def main():
     global current_dir
 
     # Clean the test_dir for residual files
-    clean_directory(test_dir)
+    helper.clean_directory(test_dir)
 
     if not os.path.isdir(test_dir):
         os.makedirs(test_dir)
@@ -226,8 +53,8 @@ def main():
     print(f"\nSwitching to {test_dir}")
 
     # Copy the scripts
-    copy_dir_files("../active_user_input", current_dir)
-    copy_dir_files("../passive_user_input", current_dir)
+    helper.copy_dir_files("../active_user_input", current_dir)
+    helper.copy_dir_files("../passive_user_input", current_dir)
 
     # select language
     language = language_input()
@@ -241,9 +68,9 @@ def main():
 
     else:
         print("Work in progress")
-
     sys.exit()
 
+'''
     # --->
     focus_area_count = 0
     metric_count = 0
@@ -408,7 +235,7 @@ def main():
 
     helper.print_summary(len(included_wgs), focus_area_count, metric_count)
 
-
+'''
   # <---
